requirements.txt
Flask==2.3.2
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.0.4
PyJWT==2.7.0
psycopg2-binary>=2.9.6   
cryptography==41.0.1        
python-dotenv==1.0.0        
alembic==1.11.1            
numpy==2.1.3
sqlalchemy==2.0.23
gunicorn==23.0.0